requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
python-dotenv
django-filter
gunicorn==20.1.0
psycopg2-binary==2.8.6
